+++ test/fast_package_test.dart
@@ -3,12 +3,13 @@ import 'fast_throttle_test/fast_throttle_test.dart' as fast_throttle_test;
 import 'fast_rate_limit_test/fast_rate_limit_test.dart' as fast_rate_limit_test;
 import 'fast_string_extension_test/fast_string_extension_test.dart'
     as fast_string_extension_test;
+import 'fast_num_extension_test/fast_num_extension_test.dart'
+    as fast_num_extension_test;
 
 void main() {
   fast_debounce_test.main();
   fast_throttle_test.main();
   fast_rate_limit_test.main();
-
-  ///
   fast_string_extension_test.main();
+  fast_num_extension_test.main();
 }